20250914 develop -> main merge 버그 및 수정사항 반영

+++ backend/src/main/java/com/trend_now/backend/board/dto/BoardInfoDto.java
@@ -12,7 +12,7 @@ public class BoardInfoDto {
     private Long boardId;
     private String boardName;
     private long boardLiveTime;
-    private Double boardExpiredTime;
+    private long boardExpiredTime;
     private double score;
     private String summary;
 }
